requirements.txt
amqp==5.0.6
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
boto3==1.20.51
botocore==1.23.51
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cycler==0.10.0
distlib==0.3.3
Django==3.2.6
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-filter==2.4.0
django-phonenumber-field==6.0.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djangorestframework-word-filter==0.2.0
filelock==3.1.0
flower==1.0.0
humanize==3.13.1
idna==3.2
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
kombu==5.1.0
matplotlib==3.4.2
mysql-connector-python==8.0.28
numpy==1.20.3
pandas==1.2.4
phonenumbers==8.12.42
Pillow==8.2.0
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
pyfcm==1.5.4
PyJWT==2.1.0
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.19.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
s3transfer==0.5.1
scikit-learn==0.24.2
scipy==1.6.3
selenium==3.141.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
stripe==2.60.0
threadpoolctl==2.1.0
tornado==6.1
urllib3==1.26.6
vine==5.0.0
virtualenv==20.8.1
wcwidth==0.2.5
webdriver-manager==3.4.2
